requirements.txt
python-dotenv==1.0.1
ipywidgets==8.1.5
google-generativeai==0.7.2
langchain==0.3.1
langchain-google-genai==2.0.0
langchain-community==0.3.1

# external agents
google-search-results==2.4.2

# LLMMathChain requires the numexpr package
numexpr==2.10.1

# Store embeddings
chromadb==0.5.18